# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventSpot API Documentation",
        "description": "API Documentation for EventSpot, a platform to discover and manage events.",
        "contact": {
            "email": "admin@eventspot.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "EventSpot API Server"
        }
    ],
    "paths": {
...
        "/api/search/user/{name}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Search users by name or email",
                "description": "Search for users whose name or email contains the specified string (case-insensitive). Returns paginated results.",
                "operationId": "d6dee3a5ae4e295645b4ecc8bfd3000e",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name or email substring to search for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users found and returned in paginated format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No users found matching the search criteria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user details by ID",
                "description": "Retrieve detailed information of a user including their interests.",
                "operationId": "d8f696420bea235f4916124c9ba1c507",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/edit/{id}": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user and their interests",
                "operationId": "878c27c9bd8eaff6428ac2004ebbcc83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User data to update, including optional interests",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the user",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "interets": {
                                        "description": "Array of interest IDs to sync with the user",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "John Doe",
                                "interets": [
                                    1,
                                    2,
                                    3
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User successfully updated with interests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User attempted to update another user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You can not update another user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name must be a string."
                                                ],
                                                "interets": [
                                                    "The interets must be an array."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InteretsResource": {
                "title": "InteretsResource",
                "description": "Représentation d'un intérêt",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "nom": {
                        "type": "string",
                        "example": "Musique"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "UserResource",
                "description": "Représentation d'un utilisateur",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "token": {
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-09T18:00:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "example": "09 July 2025"
                    },
                    "interets": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InteretsResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}